+++ types/v56/constants.go
@@ -152,4 +152,5 @@ const (
 	LBMonitorPath    = "/loadbalancer/config/monitors/"
 	LBServerPoolPath = "/loadbalancer/config/pools/"
 	LBAppProfilePath = "/loadbalancer/config/applicationprofiles/"
+	LBAppRulePath    = "/loadbalancer/config/applicationrules/"
 )