+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added functions for refreshing, getting and update Org VDC [#206](https://github.com/vmware/go-vcloud-director/pull/206)
 * Added VDC meta data create/get/delete functions [#203](https://github.com/vmware/go-vcloud-director/pull/203)
 * Added load balancer application profile [#208](https://github.com/vmware/go-vcloud-director/pull/208)
+* Added load balancer application rule [#xx](https://github.com/vmware/go-vcloud-director/pull/xx)
 
 ## 2.2.0 (May 15, 2019)